requirements.txt
Flask
flask-cors
python-dotenv
gunicorn
gensim
mecab-python3
flask-restx
